+++ _tests/go/bug_else_if_control_flow_test.go
@@ -9,7 +9,7 @@ import (
 func TestBugElseIfControlFlow(t *testing.T) {
 	t.Parallel()
 	gen := dreegotest.Build(t, map[string]string{
-		"www/routes/get.dreego": `<server>score := 85</server>
+		"www/routes/+page.dreego": `<server>score := 85</server>
 <body>
 {#if score >= 90}
 A